Only import kaniko on linux
