+++ internal/extend/kaniko/dockerfile_applier_windows.go
@@ -1,0 +1,12 @@
+//go:build windows
+
+package kaniko
+
+import (
+	"github.com/buildpacks/lifecycle/internal/extend"
+	"github.com/buildpacks/lifecycle/log"
+)
+
+func (a *DockerfileApplier) Apply(dockerfiles []extend.Dockerfile, baseImageRef string, logger log.Logger) error {
+	return nil
+}
